requirements.txt
opencv-python
numpy
Flask
Flask-SocketIO
SpeechRecognition
PyAudio
pytesseract
Pillow